v1.0.1 - fix file permissions management when commands are issued as root
When homedns-jwt add was being called it was locking the server running as "homedns" out of accessing the file because the file was retaining root as an owner.
In order to remedy this, and make the program still flexible, when new certs or config files are generated they inherit the permissions of the parent directory.

Also previously the JWT configuration was not initializing itself if it did not exist.
Now it will create an empty file with the correct permissions if it wasn't created during install (which should be ok)