Fix config defaults issues, at soa domains support
Config defaults were not being applied correctly

Also added "dns.soa_domains" to the configuration.  This allows you to specify domains that are always authoritative.
When there is an error resolving a domain, to prevent infinite recursion we need to know if that domain can be answered by forwarding it.
if it can't be shouldn't forward it, even if we encountered an error looking up that domain.